Split Adjunct from Matchers
Matchers trait and object are in their own module now.
All other classes (except for inner classes of Matchers) are in Adjunct.
